requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
flask-session
flask-ckeditor
Werkzeug==2.3.7
reportlab==4.0.4
python-docx==0.8.11
psycopg2-binary
waitress
flask-mail
flask-csrf
flask-migrate
flask-wtf
python-dotenv
flask-socketio==5.3.6
python-socketio==5.11.0
python-dateutil==2.8.2
Pillow==10.0.0
PyYAML==6.0.1
requests==2.31.0
transformers==4.36.2
accelerate==0.22.0
PyPDF2==3.0.1
pytorch==2.1.2
tensorflow==2.12.0
huggingface-hub==0.16.0
pandas
matplotlib 
seaborn
scikit-learn
scipy
nltk
spacy
gunicorn
eventlet
